==> script/4_exploratory_extreme_es.qmd <==
---
title: "Exploring extremes values (ES and SD)"
format:
  docx:
    toc: false
    number-sections: false
execute: 
  echo: false
  warning: false
knitr:
  opts_chunk:
    out.width: "100%"
    fig.align: center
---

```{r}
# Load packages
library(tidyverse)      # Data wrangling
library(metafor)        # Meta-analysis
library(readxl)         # Read Excel files
library(extrafont)      # Additional fonts
library(cowplot)        # Plot annotation and alignment
library(patchwork)      # Combine plots
```

```{r}
# Import data from Excel
df <- read_excel("../data/Dataclean_200FST.xlsx") 

# Change date type to numeric
df <- df  |>  
  mutate(year = as.numeric(format(as.Date(df$year, format = "%d/%m/%Y"),"%Y"))) 

```


```{r}
# Calculate effect size in standardized mean difference (Hedges' g)
Efeito <- escalc(measure = "SMD", 
                 n1i = ctr_n_corr, n2i = atd_n_round, 
                 m1i = ctr_mean, m2i = atd_mean, 
                 sd1i = ctr_sd, sd2i = atd_sd, 
                 data = df, 
                 append = TRUE)
```

#### EXPLORATORY ANALYSIS OF EXTREME EFFECTS ----

```{r}
extreme_studies <- Efeito |> 
  mutate(sd = sqrt(vi)) |> 
  filter(sd >= 3 & yi >= 5)

color_mice <- "#ff9400"
color_rat <- "#ec2b2b"
color_neutral <- "grey"
```

#### How many studies has a effect size \>= 5 and sd \>=3? How much % they represent of the library?

```{r}

extreme_studies_quant <- Efeito |>  
  mutate(sd = sqrt(vi)) |> 
  filter(sd >= 3 & yi >= 5) |> 
  summarise(count = n(),
            pcent = ((count * 100) / nrow(Efeito)))

print(paste("There are", extreme_studies_quant$count, "studies with extremes ES and SD, representing", format(extreme_studies_quant$pcent, digits = 4), "% of the total number of studies."))
```

#### How many publications they represent? are they nested?

```{r}
#| fig-height: 3
extreme_pub_species <- extreme_studies |> 
  group_by(species, id) |> 
  summarise(count = n()) |> 
  arrange(desc(count)) 


extreme_pub_species |> 
  ggplot2::ggplot(aes(y = count, x = reorder(id, count), fill = as.factor(species))) + 
  geom_col() + 
  scale_fill_manual(values = c("mice" = color_mice, "rat" = color_rat)) +
  scale_y_continuous(breaks = 0:10) + 
  labs(x = "Publications' ID", y = "Number of studies per publication", fill = "Species") + 
  coord_flip() +
  theme_bw()
  
```

#### All studies from these publications present a extreme effect size?

```{r}
#| fig-height: 3
Efeito |> mutate(sd = sqrt(vi)) |> 
  filter(id %in% extreme_studies$id) |> 
  mutate(extreme_or_not = case_when(
    sd >= 3 & yi >= 5 ~ TRUE,
    .default = FALSE
  )) |> 
  group_by(species, id, extreme_or_not) |> 
  count() |> 
  filter(species == "mice") |> 
  ggplot2::ggplot(aes(y = n, x = reorder(id, n), fill = as.factor(extreme_or_not))) + 
  geom_col() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_mice), labels = c("No","Yes")) +
  scale_y_continuous(breaks = 0:15) + 
  labs(x = "Publications' ID", y = "Number of studies per publication", fill = "Extreme ES and SD", title = "Mice") + 
  coord_flip() +
  theme_bw() +
  theme(legend.position = "bottom",
        plot.title = element_text(size = 8))
```


```{r}
#| fig-height: 3
Efeito |> mutate(sd = sqrt(vi)) |> 
  filter(id %in% extreme_studies$id) |> 
  mutate(extreme_or_not = case_when(
    sd >= 3 & yi >= 5 ~ TRUE,
    .default = FALSE
  )) |> 
  group_by(species, id, extreme_or_not) |> 
  count() |> 
  filter(species == "rat") |> 
  ggplot2::ggplot(aes(y = n, x = reorder(id, n), fill = as.factor(extreme_or_not))) + 
  geom_col() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_rat), labels = c("No","Yes")) +
  scale_y_continuous(breaks = 0:15) + 
  labs(x = "Publications' ID", y = "Number of studies per publication", fill = "Extreme ES and SD", title = "Rat") +
  coord_flip() +
  theme_bw() +
  theme(legend.position = "bottom",
        plot.title = element_text(size = 8))
```

#### What are the characteristics of these studies?

## by population:

```{r}
#| fig-height: 6
Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "mice") |> 
  select(sex, strain, bioterium_lightcycle, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |>
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_mice), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Mice") +
  facet_wrap(~variable, scales = "free_y", ncol = 2, nrow = 2) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = c(0.75, 0.25),
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.y = element_text(size = 8))

Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "rat") |> 
  select(sex, strain, bioterium_lightcycle, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |>
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_rat), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Rat") +
  facet_wrap(~variable, scales = "free_y", ncol = 2, nrow = 2) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = c(0.75, 0.25),
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.y = element_text(size = 8))
```

## intervention:

```{r}
#| fig-height: 6
Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "mice") |> 
  select(atd_class, atd_type, treatment_via, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |>
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_mice), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Mice") +
  facet_wrap(~variable, scales = "free_y", ncol = 2, nrow = 2) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = c(0.75, 0.25),
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.y = element_text(size = 8))

Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "rat") |> 
  select(atd_class, atd_type, treatment_via, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |>
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_rat), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Rat") +
  scale_x_discrete(
    labels = function(x)
      str_wrap(x, width = 20)
  ) +
  scale_y_continuous(n.breaks = 10) + 
  facet_wrap(~variable, scales = "free_y", ncol = 2, nrow = 2) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = c(0.75, 0.25),
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.y = element_text(size = 8))
```

## outcome:

```{r}
#| fig-height: 6
Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "mice") |> 
  select(fst_protocol, measurement_method, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |>
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_mice), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Mice") +
  facet_wrap(~variable, scales = "free_y",  ncol = 1) +
    scale_x_discrete(
    labels = function(x)
      str_wrap(x, width = 20)
  ) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = "right",
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.y = element_text(size = 8))

Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "rat") |> 
  select(fst_protocol, measurement_method, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |>
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_rat), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Rat") +
  facet_wrap(~variable, scales = "free_y", ncol = 1) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = "right",
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.y = element_text(size = 8))

```

## validity:

```{r}
#| fig-height: 3

Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "mice") |> 
  select(rob1:rob10, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |> 
   mutate(variable = reorder(variable, as.numeric(gsub("rob", "", variable)))) |> 
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_mice), labels = c("No","Yes")) +
  scale_y_continuous() + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Mice") +
  facet_wrap(~variable, nrow = 2) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = "bottom",
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.x = element_text(size = 8))

Efeito |> 
  mutate(extreme_or_not = case_when(
    line %in% extreme_studies$line ~ TRUE,
    .default = FALSE
  )) |> 
  filter(species == "rat") |> 
  select(rob1:rob10, extreme_or_not) |> 
  gather(key = "variable", value = "value", - extreme_or_not) |> 
  mutate(variable = reorder(variable, as.numeric(gsub("rob", "", variable)))) |> 
  ggplot2::ggplot(aes(x = value, fill = as.factor(extreme_or_not))) + 
  geom_bar() +
  scale_fill_manual(values = c("FALSE" = color_neutral, "TRUE" = color_rat), labels = c("No","Yes")) +
  scale_y_continuous(n.breaks = 4) + 
  labs(x = "Publication", y = "Number of studies", fill = "Extreme ES and SD", title = "Rat") +
  facet_wrap(~variable, nrow = 2) +
  coord_flip() +
  theme_bw() +
  theme(legend.position = "bottom",
        axis.title.y = element_blank(),
        plot.title = element_text(size = 8),
        axis.text.x = element_text(size = 8))


```

##### What's the effect size from the studies that reported not doing practices to ganrantee internal quality?

```{r}
rob_no <- Efeito |> mutate(sd = sqrt(vi)) |>
    filter(
      rob1 == "No"
      |rob2 == "No"
      |rob3 == "No"
      |rob4 == "No"
      |rob5 == "No"
      |rob6 == "No"
      |rob7 == "No"
      |rob8 == "No"
      |rob9 == "No"
      |rob10 == "No") 


rob_no |> 
  summarise(min_es = min(yi),
            max_es = max(yi),
            mean_es = mean(yi),
            median_es = median(yi),
            min_sd = min(sd),
            max_sd = max(sd),
            mean_sd = mean(sd),
            median_sd = median(sd)) |> gt::gt()
```


```{r}
#| fig-height: 3
#| 
rob_no |>
  rename(`Effect size` = yi,
         `Standard deviation` = sd) |> 
  pivot_longer(cols = c(`Effect size`, `Standard deviation`), names_to = "var", values_to = "value") |> ggplot(aes(x = value)) + 
  geom_histogram(binwidth = 1) + 
  facet_wrap(~var, scales = "free_x") +
  scale_x_continuous(n.breaks = 10) + 
  labs(x = "value", y = "Number of studies") +
  theme_bw()
```


#### How much do these studies with extremes ES/SD represent by the Risk of Bias?

```{r}
tb <- Efeito |> 
  mutate(
    extreme_or_not = line %in% extreme_studies$line,
    rob_no_studies = line %in% rob_no$line
  ) |> 
  labelled::set_variable_labels(
    extreme_or_not = "Extreme ES and SD",
    rob_no_studies = "High risk of bias (unless one practice)"
  ) |> 
  gtsummary::tbl_summary(
    include = c(extreme_or_not, rob_no_studies),
    missing = "no",
    by = rob_no_studies
  ) |> 
  gtsummary::bold_labels() |> 
  gtsummary::modify_header(label ~ "**High RoB**") |> 
  gtsummary::as_gt() |>
  gt::tab_footnote(footnote = "High RoB = High risk of bias at least in one practice.") 

tb
```
#### How much does the publication that contains extremes ES/SD represent by the Risk of Bias?

```{r}

tb <- Efeito |> 
  mutate(
    extreme_or_not = line %in% extreme_studies$line,
    rob_no_studies = line %in% rob_no$line
  ) |> 
  group_by(id) |> slice(1) |>  ungroup() |> 
  labelled::set_variable_labels(
    extreme_or_not = "Extreme ES and SD",
    rob_no_studies = "High risk of bias (unless one practice)"
  ) |> 
  gtsummary::tbl_summary(
    include = c(extreme_or_not, rob_no_studies),
    missing = "no",
    by = rob_no_studies
  ) |> 
  gtsummary::bold_labels() |> 
  gtsummary::modify_header(label ~ "**High RoB**") |> 
  gtsummary::as_gt() |>
  gt::tab_footnote(footnote = "High RoB = High risk of bias at least in one practice.") 

tb
```

#### What are the characteristics from these studies with extreme ES and SD that present at least one high risk of bias?

``` {r}
Efeito |> 
  mutate(
    extreme_rob = line %in% extreme_studies$line & line %in% rob_no$line) |> 
  filter(extreme_rob == TRUE) |> 
  select(
    id, species, sex, strain, bioterium_lightcycle,
    atd_class, atd_type, treatment_via,
    fst_protocol, measurement_method) |>  gt::gt()

```
